Create corpusAnalysis
This is a work in progress. I'm actively developing it now.

While it relates directly to a specific corpus (legalEnglish, at the moment) it can be re-purposed by pointing to any folder of texts, and by doing a FIND-REPLACE on 'legal' and changing it to whatever the new corpus is about. I prefer that to generic variable names in case I want to do a multi-corpus analysis one day.

After you do the find-replace it should run exactly the same ***except*** for one thing - the custom stop word list. The custom stopwords are the document-specific tokens that you don't need (e.g. abbreviations, symbols, random letters or whatnot). My process there is an iterative one, in that I'll run the code, look at the top word freqs, and start adding things I see in that list into the custom stop words CSV file. Remember if you're making a CSV from scratch to add "word," as the first line, which will by used as the column header (it tripped me up a couple times before I started remembering).